+++ generator-core/src/main/java/org/apiaddicts/apitools/apigen/generatorcore/generator/persistence/relations/OneToOneBuilder.java
@@ -2,9 +2,11 @@ package org.apiaddicts.apitools.apigen.generatorcore.generator.persistence.relat
 
 import com.squareup.javapoet.AnnotationSpec;
 
+import javax.persistence.FetchType;
 import javax.persistence.OneToOne;
 
-import static org.apiaddicts.apitools.apigen.generatorcore.generator.common.Formats.STRING;
+import static org.apiaddicts.apitools.apigen.generatorcore.generator.common.Formats.*;
+import static org.apiaddicts.apitools.apigen.generatorcore.generator.common.Members.FETCH;
 import static org.apiaddicts.apitools.apigen.generatorcore.generator.common.Members.MAPPED_BY;
 
 public class OneToOneBuilder extends RelatedFieldBuilder {
@@ -19,6 +21,7 @@ public class OneToOneBuilder extends RelatedFieldBuilder {
     protected void initialize() {
         AnnotationSpec relationAnnotation = AnnotationSpec.builder(OneToOne.class)
                 .addMember(MAPPED_BY, STRING, mappedByFieldName)
+                .addMember(FETCH, ENUM_VALUE, FetchType.class, FetchType.LAZY.name())
                 .build();
         annotations.add(relationAnnotation);
     }