+++ generator-core/src/main/java/org/apiaddicts/apitools/apigen/generatorcore/generator/persistence/relations/OneToOneOwnerBuilder.java
@@ -2,9 +2,14 @@ package org.apiaddicts.apitools.apigen.generatorcore.generator.persistence.relat
 
 import com.squareup.javapoet.AnnotationSpec;
 
+import javax.persistence.FetchType;
 import javax.persistence.OneToOne;
 
 import java.util.List;
+
+import static org.apiaddicts.apitools.apigen.generatorcore.generator.common.Formats.ENUM_VALUE;
+import static org.apiaddicts.apitools.apigen.generatorcore.generator.common.Formats.LITERAL;
+import static org.apiaddicts.apitools.apigen.generatorcore.generator.common.Members.FETCH;
 
 public class OneToOneOwnerBuilder extends RelatedFieldBuilder {
 
@@ -16,7 +21,9 @@ public class OneToOneOwnerBuilder extends RelatedFieldBuilder {
 
     @Override
     protected void initialize() {
-        annotations.add(AnnotationSpec.builder(OneToOne.class).build());
+        annotations.add(AnnotationSpec.builder(OneToOne.class)
+                .addMember(FETCH, ENUM_VALUE, FetchType.class, FetchType.LAZY.name())
+                .build());
         for (ColumnRelation cr : columns) {
             annotations.add(joinColumn(cr.getName(), cr.getReferencedName()));
         }