+++ generator-core/src/main/java/org/apiaddicts/apitools/apigen/generatorcore/generator/web/controller/endpoints/DeleteEndpointBuilder.java
@@ -4,6 +4,7 @@ import com.squareup.javapoet.TypeName;
 import lombok.extern.slf4j.Slf4j;
 import org.apiaddicts.apitools.apigen.generatorcore.config.controller.Endpoint;
 import org.apiaddicts.apitools.apigen.generatorcore.utils.Mapping;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 
 @Slf4j
@@ -19,6 +20,11 @@ public class DeleteEndpointBuilder extends EndpointBuilder {
     @Override
     protected Class getMappingClass() {
         return DeleteMapping.class;
+    }
+
+    @Override
+    protected HttpStatus getResponseStatus() {
+        return HttpStatus.NO_CONTENT;
     }
 
     @Override